+++ packages/badge/src/Badge.story.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import Badge from './Badge';
 
 export default {
-  title: 'Packages/Badge',
+  title: 'Components/Badge',
   component: Badge,
 } as Meta<typeof Badge>;