+++ packages/side-nav/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @leafygreen-ui/side-nav
 
+## 10.0.0
+
+### Patch Changes
+
+- Updated dependencies [85d46871]
+- Updated dependencies [99e20bb9]
+  - @leafygreen-ui/lib@9.4.0
+  - @leafygreen-ui/leafygreen-provider@2.3.0
+  - @leafygreen-ui/menu@14.0.0
+  - @leafygreen-ui/typography@13.0.0
+  - @leafygreen-ui/tooltip@7.0.4
+
 ## 9.0.3
 
 ### Patch Changes